+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 
 - [#84][GL-84] - Cherry trees are not supported in the Timber and Fast Leaves Decay scenarios
 - [#83][GL-83] - The explosion-power option in the Double Dates scenario does not support floating-point values
+- [#79][GL-79] - Offline zombie is not always removed
 
 [GL-84]: https://gitlab.com/uhccore/uhccore/-/issues/84
 [GL-83]: https://gitlab.com/uhccore/uhccore/-/issues/83
+[GL-79]: https://gitlab.com/uhccore/uhccore/-/issues/79
 
 ## [1.20.5] - 2023-06-16